examples/simple_bioinformatics_project/requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.5.0
biopython>=1.80
scikit-learn>=1.0

# Visualization
matplotlib>=3.5
seaborn>=0.12

# Include development requirements
-r dev-requirements.txt